Updated Inventory Management
Updated some methods for the class inventory
