--- OOP Case Study/.idea/.name
@@ -1,1 +1,0 @@
-InventoryManagement.java
